Refactored backoffice/categories component
